Viewport scroll preparations:
* Remove offset property from buffer.
* Move diffing to Terminal so that only the part of the buffer that
  the viewport overlays is diffed.
* Use interior mutability for backend.buffer.
* buffer.merge() is now O(n) and not O(2n^2).
* protect against overflow behavior when calculating `Rect.size()`.
* do not expose the buffer directly from the terminal with `current_buffer_mut`.
* added a fill widget, like the Clear widget but a bit more general.
* don't call reset() between each call.
* decouple drawing and setting cursor position. Creating runtime issues with borrow_mut and crossterm.
  Better to use terminal.set_position()

+++ src/widgets/barchart.rs
@@ -211,7 +211,7 @@ impl<'a> Widget for BarChart<'a> {
                 chart_area.left() + i as u16 * (self.bar_width + self.bar_gap),
                 chart_area.bottom() - 1,
                 label,
-                self.bar_width as usize,
+                self.bar_width,
                 self.label_style,
             );
         }
